Move ReferenceHistoryComponent under candidate namespace
