+++ spec/components/candidate_interface/reference_history_component_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe ReferenceHistoryComponent, type: :component do
+RSpec.describe CandidateInterface::ReferenceHistoryComponent, type: :component do
   it 'renders the events of a reference history', with_audited: true do
     reference = create(:reference, :not_requested_yet, created_at: Time.zone.local(2020, 1, 1, 9))
     Timecop.freeze(reference.created_at) { reference.feedback_requested! }
